+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ Breaking changes:
   splitting the data (eg. as in the examples ani_aluminum_example\*.py, 
   ani1x_training.py, and \*SNAPExample.py) must be performed using 
   PyTorch rather than NumPy. 
+- The hippynn Database object and any functions which return this object 
+  now except only an int or a torch.Generator for the ``seed`` argument. 
 
 New Features:
 -------------